requirements.txt
streamlit==1.31.1
pymupdf==1.23.6  # fitz for PDF text extraction
tiktoken==0.6.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
groq==0.6.0
python-docx==0.8.11  # For Word document generation
fpdf==1.7.2  # For PDF generation
